corrigiendo la nota de credito

+++ database/migrations/2024_02_25_024652_create_temporalncs_table.php
@@ -15,6 +15,7 @@ return new class extends Migration
     {
         Schema::create('temporalncs', function (Blueprint $table) {
             $table->id();
+            //$table->string('serienumero');
             $table->foreignId('company_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->string('codigobarras');
@@ -23,20 +24,20 @@ return new class extends Migration
             $table->string('name');
             $table->string('um');
             $table->string('tipafeigv');
-            $table->decimal('saleprice', 10, 4);
-            $table->decimal('mtovalorunitario', 10, 4)->nullable();
-            $table->decimal('porcentajeigv', 10, 4);
-            $table->decimal('factoricbper', 10, 4);
-            $table->decimal('discount', 10, 4)->nullable();
-            $table->decimal('salepricef', 10, 4)->nullable();
+            $table->double('saleprice', 10, 4);
+            $table->double('mtovalorunitario', 10, 4)->nullable();
+            $table->double('porcentajeigv', 10, 4);
+            $table->double('factoricbper', 10, 4);
+            $table->double('discount', 10, 4)->nullable();
+            $table->double('salepricef', 10, 4)->nullable();
             $table->integer('quantity');
-            //$table->decimal('subtotal', 10, 4)->nullable();
-            $table->decimal('subtotal', 10, 4)->nullable();
-            $table->decimal('igv', 10, 4)->nullable();
-            $table->decimal('totalimpuestos', 10, 4)->nullable();
-            $table->decimal('icbper', 10, 4)->nullable();
-            $table->decimal('mtovalorventa', 10, 4)->nullable();
-            $table->decimal('mtobaseigv', 10, 4)->nullable();
+            //$table->double('subtotal', 10, 4)->nullable();
+            $table->double('subtotal', 10, 4)->nullable();
+            $table->double('igv', 10, 4)->nullable();
+            $table->double('totalimpuestos', 10, 4)->nullable();
+            $table->double('icbper', 10, 4)->nullable();
+            $table->double('mtovalorventa', 10, 4)->nullable();
+            $table->double('mtobaseigv', 10, 4)->nullable();
            // $table->text('legends')->nullable();
            $table->boolean('esbolsa')->default(false);
            //$table->text('legends')->nullable(); ira en comprobante o boletas
